Make Pens inherit from GenericPen, to share similar behavior
Some, like Highlighter and Pencil, still mostly to their own thing.  But
most of the rest are rather similar, and a few are straight copies of
GenericPen.

There's odd behavior with ocolor in the pencils and paintbrush, and the
texture selection in the mechanical pencil looks broken as well.  But
I'm leaving it as is for now.