Switch rendering over to Reportlab
Most things look exactly the same.  The Highligher looks a bit different.
Reportlab seems to not do well with bevel joins with the types of paths
we get from the Remarkable.  Since RCU looks different from the Remarkable
output, wich looks different from the Remarkable screen, I'm not so worried
about that right now.

The code assumes vector output in all the pens.  The raster output is not
yet ported over to Reportlab, and it's not clear that this is an easy, or
sensible, thing to do.  The invisible dots separating the layers in vector
output are not supported yet, so OCG separation won't work.  The annotations
from highlights are not supported yet.