Generalize code to avoid requirement of named files
The input is now a generic source.  Sample sources for the filesystem
and a zipfile are included.  The output is a file-like object (right
now, a SpooledTemporaryFile).  Internally, it no longer relies on named
temporary files.  This in turn eliminates the need for cleanup code,
since these unnamed temporary files will be closed and removed when
garbage collected.

Annoyingly, Reportlab doesn't support SpooledTemporaryFiles as an
output target.