Simplify callback system
There had been two callbacks, one to report progress and the other
to check if the operation should abort.  They were almost always
called together, so this just merges them together.  The progress
callback and raise an exception if it wants this to stop.  At the
moment there's no clean up to do, so this is really easy.  In the
future, we may need to trap exceptions and do some cleanup.