Make pencil textures into a global singleton
For some reason, it was being passed through layers of objects to
avoid making multiple instances.  It makes more sense to just have
it be a global singleton.  At the moment, we're not using it at all,
but we'll leave it in place in case we do get textured brushes
working.