Handle CropBoxes and rotations
reMarkable displays based on the CropBox (but with some bugs?), if it
exists, so we should use that preferentially over the MediaBox.  The
CropBox may not be based at (0, 0), so we'll have to shift the
annotations based on the base of it.

Also, if a PDF page has a rotation, we need to counteract that in the
annotated page, so it appears upright after the merge.  But keep in
mind that the reMarkable rotates landscape pages by 90 degress when
displaying them.

At the moment, I've broken the code that handles the margins that
appear when there's an aspect-ratio mismatch.  Hopefully that will
be fixed next.