Allow alternate colors for the pens
This adds `render()` and command line parameters to specify the colors
that should be used for the "black" and "white" pen colors.  These can be
used to, for example, have annotations in red or blue lines to distinguish
from the content being annotated.

The default colors are, of course, actual black and white.  The gray pen
is always a 50/50 mix of the black and white pens in the RGB color space.